requirements.txt
tensorflow-gpu==2.0
gin-config==0.2.1
click==7.0
tqdm==4.40.2
numpy==1.17.4
